requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
sentence-transformers==2.7.0
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.7.1
requests==2.31.0
